refactor(dashboard): extract runtime-trace domain from keeper.ts
Split KeeperRuntimeTrace* + KeeperRuntimeLens* types (27 interfaces) and
parseKeeperRuntimeTrace/fetchKeeperRuntimeTrace into keeper-runtime-trace.ts
(719 lines). keeper.ts 1828 -> 1114 lines.

Self-contained domain: only isRecord (normalize) +
fetchWithTimeout/jsonHeaders/DEFAULT_GET_TIMEOUT_MS (core) deps; field
helpers (numberField etc.) stay internal. Facade re-export preserves all
29 public symbols; consumers (`from '../api/keeper'`) unchanged.

typecheck green.
